reference architectures/RAG/requirements-query.txt
requests
psycopg2-binary
gradio